+++ src/geeTetris.cpp
@@ -64,8 +64,8 @@ int main(){
     tmanager.add(&tabLevel);    // should be 1
     tmanager.add(&tabLines);
     tmanager.add(&tabShadow);
-    tmanager.add(&tabTetris, 4);
-    tmanager.add(&tabExit, 5);
+    tmanager.add(&tabTetris);
+    tmanager.add(&tabExit);
 
     // Handle keyboard events
     bool quitApp(false);
@@ -108,14 +108,7 @@ int main(){
             }
 #ifdef TRACE_MODE
             else{
-                if (KEY_CODE_CAPTURE == car){
-
-                    TRACE_DEF("Pressed 7");
-
-                    if (keys.isPressed(MOD_SHIFT)){
-                        TRACE("Pressed 7 and shift", COLOUR_BLACK, NO_COLOUR);
-                    }
-
+                if (KEY_CODE_CAPTURE == car  && keys.isPressed(MOD_SHIFT)){
                     if (!gCapture.isSet()){
                         gCapture.install();
                     }