+++ src/shared/RTC.h
@@ -1,0 +1,17 @@
+// Dark_storm
+// https://www.planet-casio.com/Fr/forums/topic12690-1-tutoriel-maitrisez-le-temps-en-cc.html#p5
+
+#ifndef __RTC_h__
+#define __RTC_h__
+
+#ifdef __cplusplus
+extern "C" {
+#endif // #ifdef __cplusplus
+
+int RTC_getTicks(void);
+
+#ifdef __cplusplus
+}
+#endif // #ifdef __cplusplus
+
+#endif // __RTC_h__